Move ServoLayoutNode and related structs to script.
