+++ components/script/Cargo.toml
@@ -48,6 +48,7 @@ phf_macros = "0.7.13"
 plugins = {path = "../plugins"}
 profile_traits = {path = "../profile_traits"}
 rand = "0.3"
+range = {path = "../range"}
 ref_filter_map = "1.0"
 ref_slice = "1.0"
 regex = "0.1.43"
