+++ components/script/lib.rs
@@ -64,6 +64,7 @@ extern crate phf;
 #[macro_use]
 extern crate profile_traits;
 extern crate rand;
+extern crate range;
 extern crate ref_filter_map;
 extern crate ref_slice;
 extern crate regex;
@@ -94,6 +95,7 @@ pub mod document_loader;
 #[macro_use]
 pub mod dom;
 pub mod layout_interface;
+pub mod layout_wrapper;
 mod mem;
 mod network_listener;
 pub mod origin;
